create: localstorageStore.js 생성
